got version update checks working on linux
now using just curl instead of wxcurl, as I could not
get wxcurl to compile without having to manually intervene (create
the wxcurl symlink). the missing symlink would cause compilation
to fail.

+++ src/features/VersionUpdateFeatureClass.h
@@ -156,7 +156,7 @@ protected:
 
 	void BackgroundWork();
 
-	wxString GetNewVersion(const wxString& currentVersion, int& statusCode);
+	wxString GetNewVersion(const wxString& currentVersion, long& statusCode);
 
 	wxString GetLabel() const;
